==> 2. tomato_7.ipynb <==
# DOWNLOADING DATA

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np

qjuices = pd.read_csv('Dataframes/qjuices.csv') # or index_col = 0 instead index = False in 'qjuices'

qjuices['date'] = pd.to_datetime(qjuices['date'])

qjuices

qjuices[qjuices['name_taste'] == 'tomato'][qjuices['id_store'] == 7][['date', 'id_sku', 'salesvolume']].to_csv('Dataframes/tomato_distribution.csv', index = False)

# QUERYING AN APPROPRIATE DATA

tomato_7 = qjuices.query('id_store == 7')[qjuices['name_taste'] == 'tomato']\
    .drop(['year', 'id_brand', 'id_store', 'name_taste', 'size', 'sizecategory'], axis = 1)\
    .drop(['weekday', 'holidays'], axis = 1)\
    .reset_index(drop = True)

tomato_7['salesvolume'] = [sale if sale >= 0 else 0 for sale in tomato_7['salesvolume']]

# tomato_7

# SELECTING NOT ONE-TIME ITEMS

items_1 = tomato_7.groupby('id_sku')['salesvolume'].agg('sum').reset_index(False)
selected_items_1 = [i for i in items_1[items_1['salesvolume'] >= 600]['id_sku']]
tomato_7 = tomato_7[tomato_7['id_sku'].isin(selected_items_1)]

# items_2 = tomato_7[tomato_7['date'] > '2013-10-01'].groupby('id_sku')['salesvolume'].agg('sum').reset_index(False)
# selected_items_2 = [i for i in items_2[items_2['salesvolume'] >= 500]['id_sku']]
# tomato_7 = tomato_7[tomato_7['id_sku'].isin(selected_items_2)]
#
# tomato_7

# CHANGING TABLE STRUCTURE

tomato_7 = tomato_7.pivot(values = ['salesvolume', 'av_price'], index = 'date', columns = 'id_sku')

tomato_7.columns = ["_".join([pair[0], str(pair[1])]) for pair in tomato_7.columns]

tomato_7.reset_index(drop = False, inplace = True)

# tomato_7

# ADDING IS_SUPPLIED

for col in [col for col in tomato_7.columns if 'salesvolume' in col]:
    tomato_7['is_supplied_' + col.replace('salesvolume_', '')] = [1 if x >= 0 else 0 for x in tomato_7[col]]

tomato_7 = tomato_7.fillna(0)

# tomato_7

# CPI

CPI = pd.read_excel('Dataframes/CPI for Fruit juices, Perm Krai, in % to the previous month.xlsx', na_values = '?')

CPI = pd.melt(CPI, id_vars = ['Month'], var_name = 'Year', value_name = 'Value')\
    .dropna(subset = ['Value'])

CPI['Date'] = CPI['Month'] + '-' + CPI['Year'].astype('str')
CPI['Date'] = pd.to_datetime(CPI['Date'])

CPI = CPI[CPI['Year'] >= 2009]
CPI.reset_index(inplace = True, drop = True)

CPI.set_index('Date', inplace = True)
CPI.drop(['Month', 'Year'], axis = 1, inplace = True)

dates = pd.date_range(CPI.index.min() - pd.DateOffset(day=1), CPI.index.max() + pd.DateOffset(day=31), freq='D', name = 'date')
CPI = CPI.reindex(dates)\
    .fillna(100)\
    .astype('float')

def cummult(series):
    A = [1.0]
    for i in range(1, len(series)):
        A.append(A[i - 1] * series[i])
    return A

CPI['cpi'] = cummult(CPI['Value'].apply(lambda x: x / 100))

CPI.drop('Value', axis = 1, inplace = True)
CPI.reset_index(inplace = True)

# CPI[CPI['date'] == '2014-12-31']
CPI

# ADJUSTING OF INFLATION

tomato_7 = tomato_7.merge(CPI, on = 'date')

for col in [col for col in tomato_7.columns if 'av_price_' in col]:
    tomato_7[col] = tomato_7[col] / tomato_7['cpi']

tomato_7.drop('cpi', axis = 1, inplace = True)

for col in [col for col in tomato_7.columns if 'av_price_' in col]:
    tomato_7[col] = tomato_7[col].replace(0, np.NaN)\
    .fillna(tomato_7[col].replace(0, np.NaN).median())
        # .fillna(method = 'ffill')\
        # .fillna(method = 'bfill')

# tomato_7

# CALCULATING DISCOUNT DATAFRAME I

discount = pd.DataFrame(tomato_7['date'])

for col in [col for col in tomato_7.columns if 'av_price_' in col]:
    med = tomato_7[col].median()
    disc = tomato_7[col].apply(lambda x: 0 if (med - x) / med < 0.25 else (med - x) / med)
    if sum(disc) > 0.25 * 20:
        discount['is_discount_' + col.replace('av_price_', '')] = disc

discount

# SELECTING ITEMS WHICH HAVE DISCOUNT

IDs = [col.strip('is_discount_') for col in discount.columns if 'is_discount_' in col]

selected_items_2 = []
for name in ['av_price_', 'salesvolume_', 'is_supplied_']:
    selected_items_2 += [col for col in tomato_7.columns if col.replace(name, '') in IDs]

tomato_7 = tomato_7[['date'] + selected_items_2]
tomato_7

# # CALCULATING DISCOUNT DATAFRAME I
#
# discount = pd.DataFrame(tomato_7['date'])
#
# for col in [col for col in tomato_7.columns if 'av_price_' in col]:
#     disc_list = []
#
#     for i, price in enumerate(tomato_7[col]):
#
#         if i != 0:
#             margin_backward = (tomato_7[col][i - 1] - price) / tomato_7[col][i - 1]
#             margin_forward = (price - tomato_7[col][i - 1]) / tomato_7[col][i - 1]
#
#         if tomato_7['is_supplied_' + col.replace('av_price_', '')][i] == 0 or i == 0:
#             disc_list.append(0)
#
#         elif margin_backward >= 0.2:
#             disc_list.append(margin_backward)   # 1
#
#         elif disc_list[i - 1] != 0 and margin_forward <= 0.2:
#             disc_list.append(disc_list[i - 1])  # 1
#
#         elif disc_list[i - 1] != 0 and margin_forward >= 0.2:
#             disc_list.append(0)
#
#         elif disc_list[i - 1] == 0 and (abs(margin_backward) <= 0.2 or price > tomato_7[col][i - 1]):
#             disc_list.append(0)
#
#     if sum(disc_list) > 5 * 0.2:
#         discount['is_discount_' + col.replace('av_price_', '')] = disc_list
#
# discount

# CALCULATING DISCOUNT DATAFRAME II

# for id in IDs:
#     days_with_discount_list = []
#     for i, val in enumerate(discount['is_discount_' + id]):
#         if val == 0:
#             days_with_discount_list.append(0)
#         else:
#             days_with_discount_list.append(days_with_discount_list[i - 1] + 1 if i > 0 else 1)
#
#     discount['num_days_discount_' + id] = days_with_discount_list
#
# for id in IDs:
#     days_without_discount_list = []
#     for i, val in enumerate(discount['is_discount_' + id]):
#         if val == 0:
#             days_without_discount_list.append((days_without_discount_list[i - 1] + 1) if i > 0 else 1)
#         else:
#             days_without_discount_list.append(0)
#
#     discount['num_days_without_discount_' + id] = days_without_discount_list

for id in IDs:
    other_dics_list = []
    for i in range(0, len(discount['is_discount_' + id])):
        other_dics_list.append(int(sum(
            [1 if val > 0 else 0 for val in discount[[col for col in discount.columns if 'is_discount_' in col]].loc[i, :]]
        )))
    discount['num_other_discounts_' + id] = other_dics_list

discount

# ADDING DUMMIES: HOLIDAYS, WEEKDAYS, MONTH

import holidays

tomato_7['weekday'] = tomato_7['date'].apply(lambda x: x.strftime('%A'))
tomato_7['month'] = tomato_7['date'].apply(lambda x: x.strftime('%B'))

tomato_7 = pd.get_dummies(tomato_7, drop_first = False)\
    .drop(['weekday_Monday',
           'month_January'], axis = 1)

# from deep_translator import GoogleTranslator
# import functools
#
# @functools.cache
# def translator(x, source = 'russian', target = 'english'):
#     return GoogleTranslator(source = source, target = target).translate(x)
#
# tomato_7['holidays'] = tomato_7['date'].apply(lambda x: holidays.RU().get(x))\
#     .replace({None:'Regular day'})\
#     .apply(lambda x: translator(x))\
#     .apply(lambda x: x.replace("'", '_').replace(' ', '_'))
#
# tomato_7['holidays_forward'] = tomato_7['holidays'].shift(-1)
# tomato_7 = tomato_7[:-1]
#
# tomato_7.reset_index(drop = True, inplace = True)
#
# tomato_7 = pd.get_dummies(tomato_7, drop_first = False)\
#     .drop(['holidays_Regular_day',
#            'holidays_forward_Regular_day'], axis = 1)

discount = discount[discount['date'] >= min(tomato_7['date'])].reset_index(drop = True)
tomato_7

tomato_7.drop([c for c in tomato_7.columns if IDs[-2] in c], axis = 1, inplace = True)
discount.drop([c for c in discount.columns if IDs[-2] in c], axis = 1, inplace = True)

# SAVING

tomato_7.to_csv('Dataframes/tomato_7.csv', index = False)

discount.to_csv('Dataframes/tomato_discount.csv', index = False)
