==> 3. Revenue.ipynb <==
# DOWNLOADING DATA

import pandas as pd
import numpy as np
import plotly.graph_objects as go
from plotly.subplots import make_subplots
from plotly import offline
from matplotlib import pyplot as plt
import warnings
warnings.filterwarnings("ignore")

tomato_discount = pd.read_csv('Dataframes/tomato_discount.csv')
tomato_7 = pd.read_csv('Dataframes/tomato_7.csv')

tomato_7

prices = [tomato_7[col] for col in tomato_7.columns if 'price' in col]
sales = [tomato_7[col] for col in tomato_7.columns if 'sales' in col]
IDs = [col.strip('is_discount_') for col in tomato_discount.columns if 'is_discount_' in col]

def revenue(quantity, price):
        return pd.Series(np.array(quantity) * np.array(price), name = 'revenue_' + quantity.name.replace('salesvolume_', ''))

def df_revenue(sales, prices):
    revenues = []

    global revenue

    for i in range(len(sales)):
        revenues.append(revenue(sales[i], prices[i]))

    return pd.DataFrame(revenues).T

rev = df_revenue(sales = sales, prices = prices)

ids_no_zero = ['46135', '75320', '201676', '415514', '415824','419020', '468496', '468558', '471503', '489135','600761']

date_test = pd.read_csv('Results/Vector models/date_test.csv')
GBDT_test_preds = pd.read_csv('Results/Vector models/Vector GBDT/GBDT_test_preds.csv')

tomato_test = tomato_7[tomato_7['date'].isin([i[0] for i in date_test.values])]

prices_test = [tomato_test[col] for col in tomato_test.columns if 'price' in col]
sales_test = [tomato_test[col] for col in tomato_test.columns if 'sales' in col]
sales_test_GBDT = [GBDT_test_preds[i].rename(str(IDs[int(i)])) for i in GBDT_test_preds.columns]

rev_test = df_revenue(sales = sales_test, prices = prices_test)[['revenue_' + i for i in ids_no_zero]]
rev_test_GBDT = df_revenue(sales = sales_test_GBDT, prices = prices_test)[['revenue_' + i for i in ids_no_zero]]

rev_test
rev_test_GBDT

print('actual 468496 test: ', rev_test['revenue_468496'].sum())
print('GBDT 468496 test: ', rev_test_GBDT['revenue_468496'].sum())
print('ratio: ',
      round(
          abs((rev_test['revenue_468496'].sum() - rev_test_GBDT['revenue_468496'].sum()) / rev_test['revenue_468496'].sum()) * 100
          , 2))

print('actual SUM test: ', rev_test.sum().sum())
print('GBDT SUM test: ', rev_test_GBDT.sum().sum())
print('ratio: ', round(abs((rev_test.sum().sum() - rev_test_GBDT.sum().sum()) / rev_test.sum().sum()) * 100, 2))

rev_sup_disc = []
rev_mean = []
rev_mean_ratio = []
for id in IDs:
    rev_sup_disc.append(pd.concat([tomato_discount['is_discount_' + id], rev[IDs.index(id)]], axis = 1)[tomato_7['is_supplied_' + id] == 1])
    rev_mean.append(rev_sup_disc[IDs.index(id)].groupby(rev_sup_disc[IDs.index(id)]['is_discount_' + id] > 0).mean())
    rev_mean_ratio.append(round(rev_mean[IDs.index(id)].iloc[1, 1] / rev_mean[IDs.index(id)].iloc[0, 1], 4))

print(f'MEAN: {np.mean(rev_mean_ratio)}',
      f'STD: {np.std(rev_mean_ratio)}',
      f'MAX: {np.max(rev_mean_ratio)}',
      f'MIN: {np.min(rev_mean_ratio)}')
rev_mean_ratio
